:wheelchair: (Score: Structure improved.) Created a ScoreForm.
migrating jsx code in ScoreField to ScoreForm from for clarifying component
duties.

BREAKING CHANGE: created a component named ScoreForm.

Closes #9
